Batch Revoke Performance Fix
Removed the useage of ProductCachedSerializationModule in
EventFactory while serializing pools since we use the
HATEOS JSON version of Pool when serializing for Events.

Removing this serializer produces the exact same JSON
data. No need to hydrate provided products as it is
never, and has never, been included.

Remove Ent collection from pool

This is a POC that removes the entitlement collection
from the pool object. This is in hopes that we do not
get lazy loading to occur when we do not need it.

NOTE: The unit tests are not yet compiling in this
branch, but the spec tests are all passing.