Round 3 of ueber data cleanup
The assumption that a Pool would never exist
without its corresponding Product was incorrect.
The combonation of customer work-arounds and the
concurrency issues while generating ueber certs
allow for Pools without its Products. The existing
cleanup queries miss this case.

This new changeset will use the Pool.productname
for the query comparisons so that they will not be
missed.